Read subitems if only_sub_subitems is set
