+++ eventio/__init__.py
@@ -1,1 +1,4 @@
-from .files import EventIOFile
+from .base import EventIOFile
+from . import iact
+
+__all__ = ['EventIOFile', 'iact']
